fix for what was a very elusive bug.  Here’s the scenario:
The sequence is E*I*P.  Now in this scenario, there is one live out from E* (say regL) and the prediction source P is of the type <opcode> regL, regL, regX.  Note that regL here is both a source and a destination register.

In this case, we were accidentally clearing out the valid field for regL, taking it off the live out list and as a result, we were not dumping live outs.

Change-Id: I29e7187e6698d8195c5d97ff0e1241e1a578722d